requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.1.4
djangorestframework==3.15.2
iniconfig==2.0.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
pytest==8.3.4
pytest-django==4.9.0
pytest-html==4.1.1
pytest-metadata==3.1.1
sqlparse==0.5.3
tzdata==2024.2
